pyproject.toml
[project]
name = "agent-sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair<5",
    "duckdb>=1.4.4",
    "ipykernel>=7.2.0",
    "ipython>=9.10.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.8",
    "langgraph-cli>=0.4.12",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas<3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.24.0",
    "typing-extensions>=4.15.0",
]
